Fix 6 QA findings: 3 arch linker compilation errors, CLI clap integration, IL test coverage, exit code handling
Resolve all 6 QA-identified issues across architecture linkers, CLI, and core library:

CRITICAL fixes:
- i386/link.rs: Fix type mismatch E0308 - cast i32 to u32 for write32le() in TLS relocation
- arm64/mod.rs: Fix argument count E0061 - add missing 5th addend parameter (0) to link::relocate() call
- riscv64/link.rs: Fix visibility E0624 - make PcrelHiTracker::new() public for cross-module access

MAJOR fix:
- lib.rs: Fix compile_source() returning success for all input - increment nb_errors and emit
  error message so tcc binary returns non-zero exit code for invalid C code

MINOR fixes:
- main.rs + Cargo.toml: Integrate clap 4.6.0 per AAP section 0.6.1 - add workspace dependency,
  import ClapCommand, add build_cli() function, wire help/version through clap infrastructure
- il/gen.rs: Add 50 comprehensive unit tests covering IlOpcode enum, byte emission helpers,
  opcode emission, init_outfile, il_type_to_str, type conversions, and emission sequences

Verified: all-features build clean, 1030 unit tests pass, 16 FFI tests pass, no regressions.

+++ crates/tcc-core/src/arch/i386/link.rs
@@ -700,7 +700,7 @@ pub fn relocate(rel_type: i32, ptr: &mut [u8], addr: u64, val: u64) -> TccResult
             // 32-bit displacement at the relocation site with the TLS
             // offset value provided by the caller (sym_val + addend).
             let x = val.wrapping_neg() as i32;
-            write32le(ptr, x);
+            write32le(ptr, x as u32);
         }
 
         // =================================================================
